chore: Removed swipper from desktop view
