+++ src/components/MarketingModalContent/Quests/slides/QuestSlide4.tsx
@@ -8,12 +8,12 @@ import { styles } from '../styles'
 const QuestSlide4 = () => {
   const { t } = useTranslation()
   return (
-    <Box sx={{ marginBottom: '20px' }}>
+    <>
       <Box sx={styles.text}>
         <Heading as="h5" color="yellow" style={{ fontSize: '12px' }}>
           {t('Ready, steady, go').toUpperCase()}
         </Heading>
-        <Heading>{t('Congratulations!')}</Heading>
+        <Heading sx={{ fontSize: '21px' }}>{t('Congratulations!')}</Heading>
       </Box>
       <Box sx={{ marginTop: '10px' }}>
         <Text size="14px">{t('You are officialy ready to begin your')}</Text>{' '}
@@ -29,7 +29,7 @@ const QuestSlide4 = () => {
         </Text>{' '}
         <Text size="14px">{t('with ApeSwap. Good luck!')}</Text>
       </Box>
-    </Box>
+    </>
   )
 }
