+++ src/components/MarketingModalContent/Quests/slides/index.tsx
@@ -1,6 +1,7 @@
+import React from 'react'
 import QuestSlide1 from './QuestSlide1'
 import QuestSlide2 from './QuestSlide2'
 import QuestSlide3 from './QuestSlide3'
 import QuestSlide4 from './QuestSlide4'
 
-export const QuestSlides = [QuestSlide1, QuestSlide2, QuestSlide3, QuestSlide4]
+export const QuestSlides = [<QuestSlide1 />, <QuestSlide2 />, <QuestSlide3 />, <QuestSlide4 />]
